move exit-code reporting to captain
