+++ cmd/state/main_test.go
@@ -30,10 +30,9 @@ func (suite *MainTestSuite) TestDeprecated() {
 	mock.MockDeprecated()
 
 	catcher := outputhelper.NewCatcher()
-	exitCode, commandString, err := run([]string{""}, catcher.Outputer)
+	exitCode, err := run([]string{""}, catcher.Outputer)
 	suite.Require().NoError(err)
 	suite.Require().Equal(0, exitCode, "Should exit with code 0, output: %s", catcher.CombinedOutput())
-	suite.Assert().Equal("", commandString)
 
 	if version.NumberIsProduction(constants.VersionNumber) {
 		suite.Require().Contains(catcher.CombinedOutput(), output.StripColorCodes(locale.Tr("warn_deprecation", "")[0:50]))
@@ -46,7 +45,7 @@ func (suite *MainTestSuite) TestExpired() {
 	mock.MockExpired()
 
 	catcher := outputhelper.NewCatcher()
-	exitCode, commandString, err := run([]string{""}, catcher.Outputer)
+	exitCode, err := run([]string{""}, catcher.Outputer)
 
 	if version.NumberIsProduction(constants.VersionNumber) {
 		suite.Require().Error(err)
@@ -56,7 +55,6 @@ func (suite *MainTestSuite) TestExpired() {
 		suite.Require().NoError(err)
 		suite.Require().Equal(0, exitCode, "Should exit with code 0, output: %s", catcher.CombinedOutput())
 	}
-	suite.Assert().Equal("", commandString)
 }
 
 func (suite *MainTestSuite) TestOutputer() {
